create users in subgroups

+++ groupAdmin/accounts_manager.php
@@ -16,6 +16,9 @@ if (!isset($_SESSION) || !isset($_SESSION['login']) || $_SESSION['login']['tempU
     die("Auth failed");
 }
 
+//error_reporting(E_ALL);
+//ini_set('display_errors', 1);
+
 
 function deleteUsersByPrefix($prefix) {
     global $db;
@@ -62,6 +65,62 @@ function prefixExists($prefix) {
 }
 
 
+function getPrefixes($idGroup) {
+    global $db;
+    $query = '
+        select
+            *
+        from
+            `groups_login_prefixes`
+        where
+            idGroup = :idGroup
+        order by
+            prefix';
+    $stm = $db->prepare($query);
+    $stm->execute([
+        'idGroup' => $idGroup
+    ]);
+
+    $res = [];
+    while($row = $stm->fetch()) {
+        $res[] = [
+            'ID' => $row['ID'],
+            'prefix' => $row['prefix'],
+        ];
+    }
+    return $res;
+}
+
+
+function createPrefix($idGroup, $prefix) {
+    global $db;
+    $query = '
+        insert into `groups_login_prefixes`
+            (idGroup, prefix)
+        values
+            (:idGroup, :prefix)';
+    $stm = $db->prepare($query);
+    $stm->execute([
+        'idGroup' => $idGroup,
+        'prefix' => $prefix
+    ]);
+}
+
+function deletePrefix($prefix) {
+    global $db;
+    $query = '
+        delete from
+            `groups_login_prefixes`
+        where
+            prefix = :prefix
+        limit 1';
+    $stm = $db->prepare($query);
+    $stm->execute([
+        'prefix' => $prefix
+    ]);
+}
+
+
 function getNewPrefix($prefix) {
     $user_prefix = getUserPrefix();
     if(!$user_prefix) {
@@ -79,42 +138,68 @@ try {
     $request = json_decode(file_get_contents("php://input"), true);
     $action = isset($request['action']) ? $request['action'] : null;
 
-    $prefix = isset($request['prefix']) ? trim($request['prefix']) : '';
-    if($prefix == '') {
-        throw new Exception('Empty prefix');
-    }
-    $client = new FranceIOI\LoginModuleClient\Client($config->login_module_client);
-    $manager = $client->getAccountsManager();
-
     switch($action) {
+
+        // delete accounts
         case 'create':
+            $prefix = isset($request['prefix']) ? trim($request['prefix']) : '';
+            if($prefix == '') {
+                throw new Exception('Empty prefix');
+            }
+            $client = new FranceIOI\LoginModuleClient\Client($config->login_module_client);
+            $manager = $client->getAccountsManager();
+
             if(preg_match("/^[a-z0-9-]+$/", $prefix) !== 1) {
                 throw new Exception('Prefix contain wrong character(s)');
             }
             $prefix = getNewPrefix($prefix);
             $amount = isset($request['amount']) ? (int) $request['amount'] : 0;
+            if(!$amount || $amount > 50) {
+                throw new Exception('Wrong amount of users');
+            }
             $postfix_length = isset($request['postfix_length']) ? (int) $request['postfix_length'] : 0;
             $password_length = isset($request['password_length']) ? (int) $request['password_length'] : 0;
-            $group_id = isset($request['group_id']) ? (int) $request['group_id'] : 0;
-            if(!$group_id) {
-                throw new Exception('Wrong grop id');
+
+            $groups = [];
+            if(isset($request['groups'])) {
+                $groups = explode(';', $request['groups']);
+                $groups = array_filter($groups, function($id) {
+                    $id = (int) $id;
+                    return $id > 0;
+                });
+                $groups = array_values($groups);
             }
+            if(!count($groups)) {
+                throw new Exception('Wrong groups param');
+            }
+
+            createPrefix($groups[0], $prefix);
+
+//$timers[] = round(microtime(true) * 1000);
             $res = $manager->create([
                 'prefix' => $prefix,
-                'amount' => $amount,
+                'amount' => $amount * count($groups),
                 'login_fixed' => true,
                 'postfix_length' => $postfix_length,
                 'password_length' => $password_length
             ]);
+//$timers[] = round(microtime(true) * 1000);
             if($res && $res['success']) {
-                foreach($res['data'] as $external_user) {
-                    $user_helper = new UserHelperClass($db);
-                    $user = $user_helper->createUser($external_user);
-                    $user_helper->addUserToGroup($user['idGroupSelf'], $group_id);
+                $i = 0;
+                $user_helper = new UserHelperClass($db);
+                foreach($groups as $group_id) {
+                    for($j=0; $j<$amount; $j++) {
+                        $external_user =& $res['data'][$i];
+                        $i++;
+                        $user = $user_helper->createUser($external_user);
+                        $user_helper->addUserToGroup($user['idGroupSelf'], $group_id);
+                        $external_user['algoreaGroupId'] = $group_id;
+                    }
                 }
                 $res = [
-                    'prefix' => $prefix,
-                    'accounts' => $res['data']
+                    'prefixes' => getPrefixes($groups[0]),
+                    'accounts' => $res['data'],
+//                    'timers' => $timers
                 ];
             } else {
                 throw new Exception(
@@ -122,7 +207,17 @@ try {
                 );
             }
             break;
+
+
+        // delete accounts
         case 'delete':
+            $prefix = isset($request['prefix']) ? trim($request['prefix']) : '';
+            if($prefix == '') {
+                throw new Exception('Empty prefix');
+            }
+            $client = new FranceIOI\LoginModuleClient\Client($config->login_module_client);
+            $manager = $client->getAccountsManager();
+
             $res = $manager->delete([
                 'prefix' => $prefix
             ]);
@@ -131,9 +226,19 @@ try {
                     $res && isset($res['error']) ? $res['error'] : 'Login module not responding'
                 );
             }
-            $res = null;
             deleteUsersByPrefix($prefix);
+            deletePrefix($prefix);
+            $res = null;
             break;
+
+        // get group account prefixes
+        case 'get_prefixes':
+            if(!isset($request['group_id'])) {
+                throw new Exception('Wrong group_id param');
+            }
+            $res = getPrefixes($request['group_id']);
+            break;
+
         default:
             throw new Exception('Incorrect action');
     }
